src/gpt-api/requirements.txt
fastapi==0.104.1
openai==1.2.3
python-dotenv==1.0.0
uvicorn==0.24.0.post1
